move play button row below artist metadata
